Now totally right for The Shah, other family left out
The task at hand was to just get everything to work right for The Shah. *using these substitutions in this order*, everything related to The Shah seems to work correctly for ~95% of cases. Note that some words/phrases and substitutions have changed since 2017-01-26. Also added some new stuff to cover common shah-related people and topics. All that stuff also appears to work right in about ~95% of cases.

At this point it's kinda like a mid-game jenga tower--very solid but be careful how you move things around. I found that if I change around the order of the substitutions there can be some unpleasant surprises.